# proto/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Task Management System API",
    "version": "1.0.0",
    "description": "Complete API documentation for Task Management System including User, Task, Notification, and Organization services"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/invite/accept": {
      "post": {
        "summary": "Accept an invite using token to complete registration",
        "operationId": "UserService_AcceptInvite",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userAcceptInviteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userAcceptInviteRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "userAcceptInviteRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        }
      },
      "title": "Accept invite request (used by invitee)"
    },
    "userAcceptInviteResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/userUser"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "Accept invite response"
    },
...
    "userUser": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/userUserRole"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "User message"
    },
    "userUserRole": {
      "type": "string",
      "enum": [
        "USER_ROLE_UNSPECIFIED",
        "USER_ROLE_MEMBER",
        "USER_ROLE_ADMIN"
      ],
      "default": "USER_ROLE_UNSPECIFIED",
      "title": "User roles"
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format: Bearer {token}"
    }
  }
}